Added Instagram methods. (test/to-do!)
First try to work on #1.

Demomstration of my attempt to include Instagram's BlinkFeed add-in into the module.
This *does not work* currently.

The method "d()" equals to the method "checkCompatibility()" in HTC Launcher, but it isn't called at any time (the Log messages aren't there).
This has to be related to the way BlinkFeed calls its add-ins' methods and Activities, because the method "d()" is th
e *only* place where the DEVICE_NOT_COMPATIBLE Intent is started from Instagram.

So, add-ins are a larger to-do.