Add more methods needed for storage
Commented out for now because I'm sure I don't have them right - need to
research what the system needs to work correctly.  Also - can this be
wrapped in a check to actually get these values from system and only
work if there is a SD card present and/or mounted?