Fixed getHDKBaseVersion by overriding it.
Not definitely sure whether 19.0 is a legit value for it, but it is the only default value I found Sense apps checking against.
May be changed in the future if another number is more appropriate.

This check fixes nothing that has to do with HDK method problems!
All methods complaining about getHDKBaseVersion may now complain about NoSuchMethodExceptions or similar.

Stay tuned!