Continue work on DownloadPreference
It now goes out online, queries the latest version, stores it, and can
set the summaries accordingly.

Download preference now has a refresh button.

Still need to add a listener to report when HTML queries are finished,
and *THEN* set the summary to indicate a new version is available.  This
works, but it's not perfect yet.